requirements.txt
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dash==1.16.2
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multitasking==0.0.9
numpy==1.19.5
pandas==1.1.5
plotly==4.1.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
retrying==1.3.3
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
yfinance==0.1.54
textblob==0.15.3
tweepy==3.10.0
PyYAML==5.3.1
regex==2020.10.15